requirements.txt
# ─── Core Data Science ───────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0

# ─── Machine Learning ────────────────────────────────────────────
scikit-learn==1.4.2
xgboost==2.0.3
imbalanced-learn==0.12.2

# ─── Visualization ───────────────────────────────────────────────
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.22.0

# ─── Experiment Tracking ─────────────────────────────────────────
mlflow==2.13.0

# ─── API ─────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1

# ─── Model Serialization ─────────────────────────────────────────
joblib==1.4.2

# ─── Notebook Support ────────────────────────────────────────────
jupyter==1.0.0
notebook==7.1.3
ipykernel==6.29.4

# ─── Utilities ───────────────────────────────────────────────────
python-dotenv==1.0.1
pyyaml==6.0.1
tqdm==4.66.4
